+++ src/controllers/CkeditorController.php
@@ -50,6 +50,8 @@ class CkeditorController extends Controller
 
         return $this->asJson([
             'url' => $asset->getUrl($transform, false),
+            'width' => $asset->getWidth($transform),
+            'height' => $asset->getHeight($transform),
         ]);
     }